moved drawing methods to abstract class
